pyproject.toml
[project]
name = "trendradar"
version = "6.7.0"
description = "TrendRadar - 热点新闻聚合与分析工具"
requires-python = ">=3.12"
dependencies = [
    "requests==2.33.0",
    "pytz==2026.1",
    "PyYAML==6.0.3",
    "fastmcp==2.12.5",
    "websockets==13.1",
    "feedparser==6.0.12",
    "boto3==1.42.76",
    "litellm==1.82.6",
    "json-repair==0.58.6",
    "tenacity==8.5.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "aiosqlite>=0.21.0",
    "python-multipart>=0.0.20",
]

[project.scripts]
trendradar = "trendradar.__main__:main"
trendradar-mcp = "mcp_server.server:run_server"

[dependency-groups]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["trendradar", "mcp_server"]
